add new traits
